Simplify Function and Parameters implementation
