+++ packages/python/src/components/SourceFile.tsx
@@ -8,7 +8,6 @@ import {
   useContext,
   type Children,
 } from "@alloy-js/core";
-
 import { join } from "pathe";
 import { PythonModuleScope } from "../symbols/index.js";
 import { ImportStatements } from "./ImportStatement.js";
